refactor: remove redundant page tests
